+++ libs/hasRole.js
