+++ libs/server/convertMongoToMeteor.js
