+++ libs/server/createIndex.js
